refactor: move acceptance test to provider_test.go
