--- internal/provider/dedicated_server_server_resource_test.go
@@ -1,8 +1,0 @@
-package provider
-
-import "testing"
-
-// TODO: We need to check if implementing tests only for import is applicable.
-func TestAccDedicatedServerResource(t *testing.T) {
-
-}
