+++ internal/provider/provider_test.go
@@ -1353,3 +1353,999 @@ func TestAccLoadBalancersDataSource(t *testing.T) {
 		},
 	})
 }
+
+func TestAccDedicatedServerNotificationSettingBandwidthResource(t *testing.T) {
+	t.Run("creates a notification setting bandwidth", func(t *testing.T) {
+		resource.Test(t, resource.TestCase{
+			ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+			Steps: []resource.TestStep{
+				// Create testing
+				{
+					Config: providerConfig + `
+resource "leaseweb_dedicated_server_bandwidth_notification_setting" "test" {
+    dedicated_server_id = "12345678"
+    frequency = "WEEKLY"
+    threshold = "1"
+    unit = "Gbps"
+}`,
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_bandwidth_notification_setting.test",
+							"id",
+							"12345",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_bandwidth_notification_setting.test",
+							"frequency",
+							"WEEKLY",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_bandwidth_notification_setting.test",
+							"threshold",
+							"1",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_bandwidth_notification_setting.test",
+							"unit",
+							"Gbps",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_bandwidth_notification_setting.test",
+							"dedicated_server_id",
+							"12345678",
+						),
+					),
+				},
+			},
+		})
+	})
+
+	t.Run(
+		"server id should be there in the request",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+resource "leaseweb_dedicated_server_bandwidth_notification_setting" "test" {
+    frequency = "WEEKLY"
+    threshold = "1"
+    unit = "Gbps"
+}`,
+						ExpectError: regexp.MustCompile(
+							"The argument \"dedicated_server_id\" is required, but no definition was found",
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"frequency should be one of these values 'DAILY', 'MONTHLY', 'WEEKLY'",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+	resource "leaseweb_dedicated_server_bandwidth_notification_setting" "test" {
+	   dedicated_server_id = "12345678"
+	   frequency = "WRONG"
+	   threshold = "1"
+	   unit = "Gbps"
+	}`,
+						ExpectError: regexp.MustCompile(
+							`Attribute frequency value must be one of: \["DAILY" "WEEKLY" "MONTHLY"]`,
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"threshold should be greater than 0",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+resource "leaseweb_dedicated_server_bandwidth_notification_setting" "test" {
+    dedicated_server_id = "12345678"
+    frequency = "DAILY"
+    threshold = "0"
+    unit = "Gbps"
+}`,
+						ExpectError: regexp.MustCompile(
+							"The value must be greater than 0, but got 0",
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"unit should be one of these values 'Mbps', 'Gbps'",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+	resource "leaseweb_dedicated_server_bandwidth_notification_setting" "test" {
+	   dedicated_server_id = "12345678"
+	   frequency = "DAILY"
+	   threshold = "0"
+	   unit = "Kbps"
+	}`,
+						ExpectError: regexp.MustCompile(
+							`Attribute unit value must be one of: \["Mbps" "Gbps"], got: "Kbps"`,
+						),
+					},
+				},
+			})
+		},
+	)
+}
+
+func TestAccControlPanelsDataSource(t *testing.T) {
+
+	t.Run(
+		"getting all control panels",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+	data "leaseweb_dedicated_server_control_panels" "dtest" {
+	}
+`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_control_panels.dtest",
+								"control_panels.#",
+								"8",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_control_panels.dtest",
+								"control_panels.0.id",
+								"CPANEL_PREMIER_100",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"filtering control panels by operating_system_id",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+	data "leaseweb_dedicated_server_control_panels" "dtest" {
+	    operating_system_id = "ALMALINUX_8_64BIT"
+	}
+`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_control_panels.dtest",
+								"control_panels.#",
+								"8",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_control_panels.dtest",
+								"control_panels.0.id",
+								"CPANEL_PREMIER_100",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+}
+
+func TestAccDedicatedServerCredentialDataSource(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			{
+				Config: providerConfig + `
+        data "leaseweb_dedicated_server_credential" "test" {
+          dedicated_server_id = "12345"
+          type                = "OPERATING_SYSTEM"
+          username            = "root"
+        }`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"data.leaseweb_dedicated_server_credential.test",
+						"dedicated_server_id",
+						"12345",
+					),
+					resource.TestCheckResourceAttr(
+						"data.leaseweb_dedicated_server_credential.test",
+						"type",
+						"OPERATING_SYSTEM",
+					),
+					resource.TestCheckResourceAttr(
+						"data.leaseweb_dedicated_server_credential.test",
+						"username",
+						"root",
+					),
+					resource.TestCheckResourceAttr(
+						"data.leaseweb_dedicated_server_credential.test",
+						"password",
+						"mys3cr3tp@ssw0rd",
+					),
+				),
+			},
+		},
+	})
+}
+
+func TestAccDedicatedServerCredentialResource(t *testing.T) {
+	t.Run("creates and updates a credential", func(t *testing.T) {
+		resource.Test(t, resource.TestCase{
+			ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+			Steps: []resource.TestStep{
+				// Create and Read testing
+				{
+					Config: providerConfig + `
+resource "leaseweb_dedicated_server_credential" "test" {
+	dedicated_server_id = "12345"
+   	username = "root"
+   	type = "OPERATING_SYSTEM"
+   	password = "mys3cr3tp@ssw0rd"
+}`,
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"dedicated_server_id",
+							"12345",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"username",
+							"root",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"type",
+							"OPERATING_SYSTEM",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"password",
+							"mys3cr3tp@ssw0rd",
+						),
+					),
+				},
+				// Update and Read testing
+				{
+					Config: providerConfig + `
+resource "leaseweb_dedicated_server_credential" "test" {
+	dedicated_server_id = "12345"
+   	username = "root"
+   	type = "OPERATING_SYSTEM"
+   	password = "mys3cr3tp@ssw0rd"
+}`,
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"dedicated_server_id",
+							"12345",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"username",
+							"root",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"type",
+							"OPERATING_SYSTEM",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_credential.test",
+							"password",
+							"mys3cr3tp@ssw0rd",
+						),
+					),
+				},
+				// Delete testing automatically occurs in TestCase
+			},
+		})
+	})
+
+	t.Run(
+		"type must be a valid one",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+resource "leaseweb_dedicated_server_credential" "test" {
+	dedicated_server_id = "12345"
+   	username = "root"
+   	type = "invalid"
+   	password = "mys3cr3tp@ssw0rd"
+}`,
+						ExpectError: regexp.MustCompile(
+							`Attribute type value must be one of: \["OPERATING_SYSTEM" "CONTROL_PANEL"(\s*)"REMOTE_MANAGEMENT" "RESCUE_MODE" "SWITCH" "PDU" "FIREWALL" "LOAD_BALANCER"],(\s*)got: "invalid"`,
+						),
+					},
+				},
+			})
+		},
+	)
+}
+
+func TestAccDataTrafficNotificationSettingResource(t *testing.T) {
+	t.Run("creates and updates a data traffic notification setting", func(t *testing.T) {
+		resource.Test(t, resource.TestCase{
+			ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+			Steps: []resource.TestStep{
+				// Create and Read testing
+				{
+					Config: providerConfig + `
+resource "leaseweb_dedicated_server_data_traffic_notification_setting" "test" {
+  dedicated_server_id = "145406"
+  frequency = "WEEKLY"
+  threshold = "1"
+  unit = "GB"
+}`,
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"id",
+							"12345",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"dedicated_server_id",
+							"145406",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"frequency",
+							"WEEKLY",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"threshold",
+							"1",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"unit",
+							"GB",
+						),
+					),
+				},
+				// Update and Read testing
+				{
+					Config: providerConfig + `
+resource "leaseweb_dedicated_server_data_traffic_notification_setting" "test" {
+  dedicated_server_id = "145406"
+  frequency = "WEEKLY"
+  threshold = "1"
+  unit = "GB"
+}`,
+					Check: resource.ComposeAggregateTestCheckFunc(
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"id",
+							"12345",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"dedicated_server_id",
+							"145406",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"frequency",
+							"WEEKLY",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"threshold",
+							"1",
+						),
+						resource.TestCheckResourceAttr(
+							"leaseweb_dedicated_server_data_traffic_notification_setting.test",
+							"unit",
+							"GB",
+						),
+					),
+				},
+				// Delete testing automatically occurs in TestCase
+			},
+		})
+	})
+
+	t.Run(
+		"threshold must be greater than 0",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+resource "leaseweb_dedicated_server_data_traffic_notification_setting" "test" {
+  dedicated_server_id = "145406"
+  frequency = "WEEKLY"
+  threshold = "-1"
+  unit = "GB"
+}`,
+						ExpectError: regexp.MustCompile(
+							"The value must be greater than 0, but got -1",
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"unit must be one of GB,MB,TB",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+resource "leaseweb_dedicated_server_data_traffic_notification_setting" "test" {
+  dedicated_server_id = "145406"
+  frequency = "WEEKLY"
+  threshold = "1"
+  unit = "blah"
+}`,
+						ExpectError: regexp.MustCompile(
+							`Attribute unit value must be one of: \["MB" "GB" "TB"], got: "blah"`,
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"frequency must be one of DAILY,WEEKLY,MONTHLY",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+resource "leaseweb_dedicated_server_data_traffic_notification_setting" "test" {
+  dedicated_server_id = "145406"
+  frequency = "blah"
+  threshold = "1"
+  unit = "GB"
+}`,
+						ExpectError: regexp.MustCompile(
+							`Attribute frequency value must be one of: \["DAILY" "WEEKLY" "MONTHLY"], got:`,
+						),
+					},
+				},
+			})
+		},
+	)
+}
+
+func TestAccDedicatedServerInstallationResource(t *testing.T) {
+	t.Run("install os on a dedicated server",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Create testing
+					{
+						Config: providerConfig + `
+    resource "leaseweb_dedicated_server_installation" "test" {
+      dedicated_server_id = "12345"
+      operating_system_id = "UBUNTU_22_04_64BIT"
+    }`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"leaseweb_dedicated_server_installation.test",
+								"id",
+								"bcf2bedf-8450-4b22-86a8-f30aeb3a38f9",
+							),
+							resource.TestCheckResourceAttr(
+								"leaseweb_dedicated_server_installation.test",
+								"dedicated_server_id",
+								"12345",
+							),
+							resource.TestCheckResourceAttr(
+								"leaseweb_dedicated_server_installation.test",
+								"operating_system_id",
+								"UBUNTU_22_04_64BIT",
+							),
+						),
+					},
+				},
+			})
+		})
+
+	t.Run(
+		"server id should be in the request",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+    resource "leaseweb_dedicated_server_installation" "test" {
+      operating_system_id = "UBUNTU_22_04_64BIT"
+    }`,
+						ExpectError: regexp.MustCompile(
+							"The argument \"dedicated_server_id\" is required, but no definition was found",
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"operating system id should be in the request",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+    resource "leaseweb_dedicated_server_installation" "test" {
+      dedicated_server_id = "12345"
+    }`,
+						ExpectError: regexp.MustCompile(
+							"The argument \"operating_system_id\" is required, but no definition was found",
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"raid.level should be one of these values '0', '1', '5', '10'",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+		resource "leaseweb_dedicated_server_installation" "test" {
+      dedicated_server_id = "12345"
+      operating_system_id = "UBUNTU_22_04_64BIT"
+      raid = {
+        level = 11
+      }
+    }`,
+						ExpectError: regexp.MustCompile(
+							`Attribute raid.level value must be one of: \["0" "1" "5" "10"]`,
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"raid.type should be one of these values 'HW', 'SW', 'NONE'",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+		resource "leaseweb_dedicated_server_installation" "test" {
+      dedicated_server_id = "12345"
+      operating_system_id = "UBUNTU_22_04_64BIT"
+      raid = {
+        type = "TEST"
+      }
+    }`,
+						ExpectError: regexp.MustCompile(
+							`Attribute raid.type value must be one of: \["HW" "SW" "NONE"]`,
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"ssh_keys should be set of string",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+		resource "leaseweb_dedicated_server_installation" "test" {
+      dedicated_server_id = "12345"
+      operating_system_id = "UBUNTU_22_04_64BIT"
+      ssh_keys = "test keys"
+    }`,
+						ExpectError: regexp.MustCompile(
+							`Inappropriate value for attribute "ssh_keys": set of string required`,
+						),
+					},
+				},
+			})
+		},
+	)
+}
+
+func TestAccOperatingSystemsDataSource(t *testing.T) {
+
+	t.Run(
+		"getting all operating systems",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+	data "leaseweb_dedicated_server_operating_systems" "dtest" {
+	}
+`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_operating_systems.dtest",
+								"operating_systems.#",
+								"24",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_operating_systems.dtest",
+								"operating_systems.0.id",
+								"ALMALINUX_8_64BIT",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_operating_systems.dtest",
+								"operating_systems.0.name",
+								"AlmaLinux 8 (x86_64)",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"filtering operating systems by control_panel_id",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+	data "leaseweb_dedicated_server_operating_systems" "dtest" {
+		control_panel_id = "CPANEL_PREMIER_100"
+	}
+`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_operating_systems.dtest",
+								"operating_systems.#",
+								"24",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_operating_systems.dtest",
+								"operating_systems.0.id",
+								"ALMALINUX_8_64BIT",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_operating_systems.dtest",
+								"operating_systems.0.name",
+								"AlmaLinux 8 (x86_64)",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+}
+
+func TestAccDedicatedServerDataSource(t *testing.T) {
+	t.Run(
+		"getting dedicated server detail by id",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+		data "leaseweb_dedicated_server_server" "test" {
+			id = "12345"
+		}`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"id",
+								"12345",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"asset_id",
+								"627294",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"serial_number",
+								"JDK18291JK",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"contract_id",
+								"674382",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"rack_type",
+								"DEDICATED",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"is_automation_feature_available",
+								"true",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"is_ipmi_reboot_feature_available",
+								"false",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"is_power_cycle_feature_available",
+								"true",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"is_private_network_feature_available",
+								"true",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"is_remote_management_feature_available",
+								"false",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"location_rack",
+								"13",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"location_site",
+								"AMS-01",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"location_suite",
+								"A6",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"location_unit",
+								"16-17",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"public_mac",
+								"AA:BB:CC:DD:EE:FF",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"public_ip",
+								"123.123.123.123/27",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"public_gateway",
+								"123.123.123.126",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"internal_mac",
+								"AA:BB:CC:DD:EE:FF",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"internal_ip",
+								"123.123.123.123/27",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"internal_gateway",
+								"123.123.123.126",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"ram_size",
+								"32",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"ram_unit",
+								"GB",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"cpu_quantity",
+								"4",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_server.test",
+								"cpu_type",
+								"Intel Xeon E3-1220",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"id is required for getting the dedicated server detail",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					{
+						Config: providerConfig + `
+		data "leaseweb_dedicated_server_server" "test" {
+		}`,
+						ExpectError: regexp.MustCompile(
+							`The argument "id" is required, but no definition was found`,
+						),
+					},
+				},
+			})
+		},
+	)
+}
+
+func TestAccDedicatedServersDataSource(t *testing.T) {
+	t.Run(
+		"getting dedicated servers by reference",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+		data "leaseweb_dedicated_server_servers" "test" {
+			reference = "test-reference"
+		}`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.test",
+								"ids.#",
+								"2",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.test",
+								"reference",
+								"test-reference",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"getting dedicated servers",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+		data "leaseweb_dedicated_server_servers" "test" {
+		}`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.test",
+								"ids.#",
+								"2",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+
+	t.Run(
+		"getting dedicated servers with all filters",
+		func(t *testing.T) {
+			resource.Test(t, resource.TestCase{
+				ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+				Steps: []resource.TestStep{
+					// Read testing
+					{
+						Config: providerConfig + `
+		data "leaseweb_dedicated_server_servers" "filter" {
+			reference = "test-reference"
+			ip = "127.0.0.4"
+			mac_address = "aa:bb:cc:dd:ee:ff"
+			site = "ams-01"
+			private_rack_id = "r id"
+			private_network_capable = "true"
+			private_network_enabled = "true"
+		}`,
+						Check: resource.ComposeAggregateTestCheckFunc(
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"ids.#",
+								"2",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"reference",
+								"test-reference",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"ip",
+								"127.0.0.4",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"mac_address",
+								"aa:bb:cc:dd:ee:ff",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"site",
+								"ams-01",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"private_rack_id",
+								"r id",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"private_network_capable",
+								"true",
+							),
+							resource.TestCheckResourceAttr(
+								"data.leaseweb_dedicated_server_servers.filter",
+								"private_network_enabled",
+								"true",
+							),
+						),
+					},
+				},
+			})
+		},
+	)
+
+}
